collectCropReward event
